special disparition done

+++ groupe1.info3.pla.game/game.purgatoire/edu/ricm3/game/purgatoire/Timer.java
@@ -39,7 +39,7 @@ public class Timer {
 		m_duration = duration;
 	}
 
-	public double getRemainingTimePercentage() {
+	public float getRemainingTimePercentage() {
 		return (float) m_current / m_duration;
 	}
